pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "speech-to-text"
version = "0.1.0"
description = "Continuous speech-to-text library with pluggable engines (Vosk, Whisper, Deepgram)"
requires-python = ">=3.10"
license = "MIT"
readme = "README.md"
keywords = ["speech", "stt", "speech-to-text", "vosk", "whisper", "deepgram", "transcription"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]
dependencies = [
    "sounddevice>=0.4",
    "numpy>=1.24",
]

[project.optional-dependencies]
vosk = ["vosk>=0.3.45"]
whisper = ["faster-whisper>=1.0.0"]
deepgram = ["deepgram-sdk>=3.0.0"]
all = [
    "vosk>=0.3.45",
    "faster-whisper>=1.0.0",
    "deepgram-sdk>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/Ch3vg/SpeechToText"
Repository = "https://github.com/Ch3vg/SpeechToText"
Issues = "https://github.com/Ch3vg/SpeechToText/issues"

[tool.setuptools.packages.find]
include = ["speech_to_text*"]
